refactor: Move components to MailCoreUI
